Code to create folder on shared machine is working

+++ MiniProjects/ExcelFileCreation/src/main/java/com/excel/create/ExcelFileCreationApplication.java
@@ -16,11 +16,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import com.excel.create.service.ExcelFileService;
 
 @SpringBootApplication
-public class ExcelFileCreationApplication implements CommandLineRunner{
+public class ExcelFileCreationApplication implements CommandLineRunner {
 
 	@Autowired
 	ExcelFileService excelFileService;
-	
+
 	public static void main(String[] args) {
 		SpringApplication.run(ExcelFileCreationApplication.class, args);
 	}
@@ -35,30 +35,36 @@ public class ExcelFileCreationApplication implements CommandLineRunner{
 //			System.out.println("File is not present and hence creating it.....");
 //			excelFileService.createExcelFile();
 //		}
-		
-		
-		String networkFilePath = "\\\\ASHWINHSACODES\\SharedFolder\\Orders.xls";
 
-        try {
-            // Create a File object representing the file on the network drive
-            File networkFile = new File(networkFilePath);
+		Path filePath = Paths.get("\\\\\\\\ASHWINS\\\\SharedFolder\\\\Orders.xls");
 
-            // Create the new file if it doesn't already exist
-            if (networkFile.createNewFile()) {
-                System.out.println("File created successfully on network drive: " + networkFile.getAbsolutePath());
+		if (Files.exists(filePath)) {
+			System.out.println("File is present....");
+		} else {
+			System.out.println("File is not present and hence creating it.....");
+			String networkFilePath = "\\\\ASHWINS\\SharedFolder\\Orders.xls";
 
-                // Optional: Write content to the file
-                try (FileWriter writer = new FileWriter(networkFile)) {
-                    writer.write("This is some content for the network file.");
-                    System.out.println("Content written to the file.");
-                }
-            } else {
-                System.out.println("File already exists on network drive: " + networkFile.getAbsolutePath());
-            }
+			try {
+				// Create a File object representing the file on the network drive
+				File networkFile = new File(networkFilePath);
 
-        } catch (IOException e) {
-            System.err.println("An error occurred while creating or writing to the file on the network drive:");
-            e.printStackTrace();
-        }
+				// Create the new file if it doesn't already exist
+				if (networkFile.createNewFile()) {
+					System.out.println("File created successfully on network drive: " + networkFile.getAbsolutePath());
+
+					// Optional: Write content to the file
+					try (FileWriter writer = new FileWriter(networkFile)) {
+						writer.write("This is some content for the network file.");
+						System.out.println("Content written to the file.");
+					}
+				} else {
+					System.out.println("File already exists on network drive: " + networkFile.getAbsolutePath());
+				}
+
+			} catch (IOException e) {
+				System.err.println("An error occurred while creating or writing to the file on the network drive:");
+				e.printStackTrace();
+			}
+		}
 	}
 }
